mise en place des autorisation des user

+++ app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Role;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -66,6 +67,11 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        // si l'utilisateur n'est pas un admin retour la page d'accuile
+        if (Gate::denies('edit-users')) {
+
+           return redirect()->route('admin.users.index');
+        }
 
         $roles = Role::all();
 
@@ -99,6 +105,13 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        /* verificaton du droit de supression en et utilisant le gate dans le provider avec */
+         
+        if (Gate::denies('delete-users')) {
+
+           return redirect()->route('admin.users.index');
+        }
+
         $user->roles()->detach();
         $user->delete();
